requirements.txt
# Web Framework
Flask==3.0.0
Werkzeug==3.0.1

# Data Processing
pandas==2.1.4
openpyxl==3.1.2

# Database
SQLAlchemy==2.0.23
PyMySQL==1.1.0
cryptography==41.0.7

# SSH Tunneling
sshtunnel==0.4.0
paramiko==3.4.0

# Utilities
python-dotenv==1.0.0

# Production Server (optional)
gunicorn==21.2.0

